adc emulation works hooray
